add bookmarks storage service
